torch-infer/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
torch>=2.5.0
torchvision>=0.20.0
transformers>=4.47.0
Pillow>=11.0.0
httpx>=0.28.0
pydantic>=2.10.0
